Add support for basic auth endpoints (e.g. JWT tokens)
There are a few endpoints in the PrimeTrust API that don't follow their
normal RESTful model, like /auth/jwts, so this adds a method basic_req
to PrimeTrust.API that allows a request to be made to those endpoints.

This also adds a helper to generate a new JWT from a user's credentials.